Generate unique detail section component for final screens and standardize tests
