+++ Example/KhipuClientIOS/ViewController.swift
@@ -18,7 +18,7 @@ class ViewController: UIViewController {
 
 
 
-        sampleTextField.text = "jlzy1kxhtzw8"
+        sampleTextField.text = "bglpckoc9irz"
         sampleTextField.font = UIFont.systemFont(ofSize: 15)
         sampleTextField.borderStyle = UITextField.BorderStyle.roundedRect
         sampleTextField.autocorrectionType = UITextAutocorrectionType.no
