+++ KhipuClientIOS/Classes/SwiftUiClient/Components/FailureMessageComponent.swift
@@ -43,7 +43,8 @@ struct FailureMessageComponent: View {
             
             FormWarning(text: operationFailure.body ?? "")
             Spacer().frame(height:Dimens.Spacing.large)
-            DetailSectionFailure(operationFailure: operationFailure,operationInfo: viewModel.uiState.operationInfo, viewModel: viewModel)
+            
+            DetailSectionComponent(reason: FieldUtils.getFailureReasonCode(reason: operationFailure.reason), operationId: operationFailure.operationID!,operationInfo: viewModel.uiState.operationInfo,viewModel: viewModel)
             
             MainButton(
                 text: viewModel.uiState.translator.t("default.end.and.go.back"),
@@ -63,57 +64,6 @@ struct FailureMessageComponent: View {
     
 }
 
-@available(iOS 15.0.0, *)
-struct DetailSectionFailure: View {
-    var operationFailure: OperationFailure
-    var operationInfo: OperationInfo?
-    @ObservedObject public var viewModel: KhipuViewModel
-    @EnvironmentObject private var themeManager: ThemeManager
-    
-    var body: some View {
-        
-        VStack(alignment: .center, spacing:Dimens.Spacing.large) {
-            Text(viewModel.uiState.translator.t("default.detail.label"))
-                .font(themeManager.selectedTheme.fonts.font(style: .semiBold, size: 16))
-            
-            DetailItemFailure(label: viewModel.uiState.translator.t("default.amount.label"), value: operationInfo?.amount ?? "")
-            
-            DetailItemFailure(label: viewModel.uiState.translator.t("default.merchant.label"), value:operationInfo?.merchant?.name ?? "")
-            DashedLine()
-            DetailItemFailure(label: viewModel.uiState.translator.t("default.operation.code.short.label"), value: FieldUtils.formatOperationId(operationId: operationFailure.operationID)+" "+FieldUtils.getFailureReasonCode(reason: operationFailure.reason),shouldCopyValue: true)
-        }
-        .padding(Dimens.Padding.large)
-        .frame(maxWidth: .infinity, alignment: .top)
-        .cornerRadius(Dimens.CornerRadius.moderatelySmall)
-    }
-}
-
-@available(iOS 15.0.0, *)
-struct DetailItemFailure: View {
-    var label: String
-    var value: String
-    var shouldCopyValue: Bool = false
-    @EnvironmentObject private var themeManager: ThemeManager
-    
-    var body: some View {
-        HStack {
-            Text(label)
-                .font(themeManager.selectedTheme.fonts.font(style: .medium, size: 14))
-                .foregroundColor(themeManager.selectedTheme.colors.labelForeground)
-            Spacer()
-            if !shouldCopyValue {
-                Text(value)
-                    .font(themeManager.selectedTheme.fonts.font(style: .semiBold, size: 14))
-                
-            } else {
-                CopyToClipboardOperationId(text: value, textToCopy: FieldUtils.formatOperationId(operationId:value), background:themeManager.selectedTheme.colors.onSecondaryContainer)
-            }
-        }
-        .padding(.vertical,Dimens.Padding.verySmall)
-    }
-}
-
-
 @available(iOS 15.0, *)
 struct FailureMessageComponent_Previews: PreviewProvider {
     static var previews: some View {
@@ -129,39 +79,6 @@ struct FailureMessageComponent_Previews: PreviewProvider {
                     title: "Title",
                     reason: FailureReasonType.taskExecutionError
                 ), viewModel: KhipuViewModel())
-        .environmentObject(ThemeManager())
-        .padding()
-    }
-}
-
-@available(iOS 15.0, *)
-struct DetailSectionFailure_Previews: PreviewProvider {
-    static var previews: some View {
-        DetailSectionFailure(
-            operationFailure:
-                OperationFailure(
-                    type: MessageType.operationFailure,
-                    body: "body",
-                    events: nil,
-                    exitURL: "exitUrl",
-                    operationID: "operationID",
-                    resultMessage: "resultMessage",
-                    title: "Title",
-                    reason: FailureReasonType.taskExecutionError
-                ), viewModel: KhipuViewModel())
-        .environmentObject(ThemeManager())
-        .padding()
-    }
-}
-
-
-@available(iOS 15.0, *)
-struct DetailItemFailure_Previews: PreviewProvider {
-    static var previews: some View {
-        DetailItemFailure(
-            label: "Label",
-            value: "Value"
-        )
         .environmentObject(ThemeManager())
         .padding()
     }
