+++ KhipuClientIOS/Classes/SwiftUiClient/Components/WarningMessageComponent.swift
@@ -45,7 +45,9 @@ struct WarningMessageComponent: View {
             
             FormWarning(text: operationWarning.body ?? "")
             Spacer().frame(height:Dimens.Spacing.large)
-            DetailSectionWarning(operationWarning: operationWarning,operationInfo: viewModel.uiState.operationInfo, viewModel: viewModel)
+            
+            
+            DetailSectionComponent(reason: FieldUtils.getFailureReasonCode(reason: operationWarning.reason), operationId: operationWarning.operationID!,operationInfo: viewModel.uiState.operationInfo,viewModel: viewModel)
             
             MainButton(
                 text: viewModel.uiState.translator.t("default.end.and.go.back"),
@@ -63,55 +65,6 @@ struct WarningMessageComponent: View {
     }
 }
 
-@available(iOS 15.0.0, *)
-struct DetailSectionWarning: View {
-    var operationWarning: OperationWarning
-    var operationInfo: OperationInfo?
-    @ObservedObject public var viewModel: KhipuViewModel
-    @EnvironmentObject private var themeManager: ThemeManager
-    
-    var body: some View {
-        VStack(alignment: .center, spacing:Dimens.Spacing.large) {
-            Text(viewModel.uiState.translator.t("default.detail.label"))
-                .font(themeManager.selectedTheme.fonts.font(style: .semiBold, size: 16))
-            
-            DetailItemWarning(label: viewModel.uiState.translator.t("default.amount.label"), value: operationInfo?.amount ?? "")
-            
-            DetailItemWarning(label: viewModel.uiState.translator.t("default.merchant.label"), value:operationInfo?.merchant?.name ?? "")
-            DashedLine()
-            DetailItemWarning(label: viewModel.uiState.translator.t("default.operation.code.short.label"), value: FieldUtils.formatOperationId(operationId: operationWarning.operationID)+" "+FieldUtils.getFailureReasonCode(reason: operationWarning.reason),shouldCopyValue: true)
-        }
-        .padding(20)
-        .frame(maxWidth: .infinity, alignment: .top)
-        .cornerRadius(6)
-    }
-}
-
-@available(iOS 15.0.0, *)
-struct DetailItemWarning: View {
-    var label: String
-    var value: String
-    var shouldCopyValue: Bool = false
-    @EnvironmentObject private var themeManager: ThemeManager
-    
-    var body: some View {
-        HStack {
-            Text(label)
-                .font(themeManager.selectedTheme.fonts.font(style: .medium, size: 14))
-                .foregroundColor(themeManager.selectedTheme.colors.labelForeground)
-            Spacer()
-            if !shouldCopyValue {
-                Text(value)
-                    .font(themeManager.selectedTheme.fonts.font(style: .semiBold, size: 14))
-                
-            } else {
-                CopyToClipboardOperationId(text: value, textToCopy: FieldUtils.formatOperationId(operationId:value), background:themeManager.selectedTheme.colors.onSecondaryContainer)
-            }
-        }
-        .padding(.vertical,Dimens.Padding.verySmall)
-    }
-}
-
 @available(iOS 15.0, *)
 struct WarningMessageComponent_Previews:PreviewProvider{
     static var previews: some View{
@@ -126,50 +79,6 @@ struct WarningMessageComponent_Previews:PreviewProvider{
                                             title: "Title",
                                             reason: FailureReasonType.taskDumped
                                         ), viewModel: KhipuViewModel()
-        )
-        .environmentObject(ThemeManager())
-        .padding()
-    }
-}
-
-@available(iOS 15.0, *)
-struct DetailSectionWarning_Previews:PreviewProvider{
-    static var previews: some View{
-        return DetailSectionWarning(
-            operationWarning: OperationWarning(
-                type: MessageType.operationWarning,
-                body: "body",
-                events: nil,
-                exitURL: "exitUrl",
-                operationID: "operationID",
-                resultMessage: "resultMessage",
-                title: "Title",
-                reason: FailureReasonType.taskDumped
-            ), operationInfo: OperationInfo(
-                acceptManualTransfer: true,
-                amount: "$ 1.000",
-                body: "body",
-                email: "khipu@khipu.com",
-                merchant: nil,
-                operationID: "operationID",
-                subject: "Subject",
-                type: MessageType.operationInfo,
-                urls: nil,
-                welcomeScreen: nil
-            ), viewModel: KhipuViewModel()
-        )
-        .environmentObject(ThemeManager())
-        .padding()
-    }
-}
-
-@available(iOS 15.0, *)
-struct DetailItemWarning_Previews:PreviewProvider{
-    static var previews: some View{
-        return DetailItemWarning(
-            label: "Label",
-            value: "Value",
-            shouldCopyValue: true
         )
         .environmentObject(ThemeManager())
         .padding()
